// src/types/order.ts
export interface Order {
  id: string;
  codigo_pedido: string | null;
  codigo_entrega?: string | null;
  nome_cliente: string | null;
  telefone: string | null;
  endereco_completo: string | null;
  bairro: string | null;
  itens: string | null;
  valor_total: string | null;
  valor_pedido: string | null;
  taxa_entrega: string | null;
  forma_pagamento: string | null;
  tempo_entrega_estimado: string | null;
  entregador?: string | null;
  entregador_usuario?: string | null;
  codigo_validado_em?: string | null;
  status: string | null;
  data_hora_pedido: string;
  criado_em: string | null;
}